+++ src/main/kotlin/com/denishovart/freqs/playlist/dto/TrackInput.kt
@@ -9,7 +9,7 @@ data class TrackInput(
     val name: String,
     val albumName: String,
     val image: String?,
-    val duration: Float,
+    val duration: Int,
     val spotifyId: String,
     val spotifyAlbumId: String,
     val artists: List<Artist>
